+++ apps/prairielearn/elements/pl-drawing/pl-drawing.js
@@ -56,7 +56,7 @@ window.PLDrawingApi = {
    * Generates a new numeric ID for a submission object.
    * Each submitted object is uniquely identified by its ID.
    */
-  generateID: function () {
+  generateID() {
     return this._idCounter++;
   },
 
@@ -69,7 +69,7 @@ window.PLDrawingApi = {
    * @param extensionName Name of the extension/group of elements.
    * @param dictionary Dictionary of elements to register.
    */
-  registerElements: function (extensionName, dictionary) {
+  registerElements(extensionName, dictionary) {
     _.extend(this.elements, dictionary);
     Object.keys(dictionary).forEach((elem) => {
       this.elementModule[elem] = extensionName;
@@ -82,7 +82,7 @@ window.PLDrawingApi = {
    * @param options Element options.  Must contain a 'type' key.
    * @param submittedAnswer Answer state object.
    */
-  createElement: function (canvas, options, submittedAnswer) {
+  createElement(canvas, options, submittedAnswer) {
     const name = options.type;
     let added = null;
 
@@ -101,7 +101,7 @@ window.PLDrawingApi = {
    * @param name Name of the element to look up.
    * @returns The element, if found.  Silently fails with the base element otherwise.
    */
-  getElement: function (name) {
+  getElement(name) {
     let ret = PLDrawingBaseElement;
     if (name in this.elements) {
       let elem = this.elements[name];
@@ -117,7 +117,7 @@ window.PLDrawingApi = {
    * @param canvas Canvas to restore state onto.
    * @param submittedAnswer Answer state to restore from.
    */
-  restoreAnswer: function (canvas, submittedAnswer) {
+  restoreAnswer(canvas, submittedAnswer) {
     for (const [id, obj] of Object.entries(submittedAnswer._answerData)) {
       this._idCounter = Math.max(parseInt(id) + 1, this._idCounter);
       let newObj = JSON.parse(JSON.stringify(obj));
@@ -132,7 +132,7 @@ window.PLDrawingApi = {
    * @param elem_options Any options to give to the element
    * @param existing_answer_submission Existing submission to place on the canvas.
    */
-  setupCanvas: function (root_elem, elem_options, existing_answer_submission) {
+  setupCanvas(root_elem, elem_options, existing_answer_submission) {
     let canvas_elem = $(root_elem).find('canvas')[0];
     let canvas_width = parseFloat(elem_options.width);
     let canvas_height = parseFloat(elem_options.height);
@@ -248,7 +248,7 @@ window.PLDrawingApi = {
     fabric.util.addListener(canvas.upperCanvasEl, 'dblclick', function (e) {
       const target = canvas.findTarget(e);
       if (target !== undefined) {
-        target.fire('dblclick', { e: e });
+        target.fire('dblclick', { e });
       }
     });