+++ apps/prairielearn/src/api/v1/endpoints/courseInstanceInfo/index.js
@@ -1,12 +1,12 @@
 // @ts-check
 const asyncHandler = require('express-async-handler');
-const path = require('path');
-const express = require('express');
-const router = express.Router({ mergeParams: true });
+import * as path from 'node:path';
+import { Router } from 'express';
 
-const sqldb = require('@prairielearn/postgres');
+import * as sqldb from '@prairielearn/postgres';
 
 const sql = sqldb.loadSql(path.join(__dirname, '..', 'queries.sql'));
+const router = Router({ mergeParams: true });
 
 router.get(
   '/',
@@ -18,4 +18,4 @@ router.get(
   }),
 );
 
-module.exports = router;
+export default router;