+++ apps/prairielearn/src/ee/auth/saml/router.ts
@@ -59,7 +59,7 @@ router.post(
     const institutionId = req.params.institution_id;
     const institutionSamlProvider = await getInstitutionSamlProvider(institutionId);
     if (!institutionSamlProvider) {
-      throw error.make(404, 'Institution does not support SAML authentication');
+      throw new error.HttpStatusError(404, 'Institution does not support SAML authentication');
     }
 
     const uidAttribute = institutionSamlProvider.uid_attribute;
@@ -115,7 +115,7 @@ router.get(
   asyncHandler(async (req, res, next) => {
     const samlProvider = await getInstitutionSamlProvider(req.params.institution_id);
     if (!samlProvider) {
-      throw error.make(404, 'Institution does not support SAML authentication');
+      throw new error.HttpStatusError(404, 'Institution does not support SAML authentication');
     }
 
     strategy.generateServiceProviderMetadata(