+++ apps/prairielearn/src/lib/db-types.ts
@@ -84,6 +84,9 @@ export const DateFromISOString = z
   )
   .transform((s) => new Date(s));
 
+export const ModeSchema = z.enum(['Public', 'Exam', 'SEB']);
+export type Mode = z.infer<typeof ModeSchema>;
+
 export const CourseSchema = z.object({
   branch: z.string(),
   commit_hash: z.string().nullable(),
@@ -150,6 +153,9 @@ export const SamlProviderSchema = z.object({
   sso_login_url: z.string(),
   uid_attribute: z.string().nullable(),
   uin_attribute: z.string().nullable(),
+  validate_audience: z.boolean(),
+  want_assertions_signed: z.boolean(),
+  want_authn_response_signed: z.boolean(),
 });
 export type SamlProvider = z.infer<typeof SamlProviderSchema>;
 
@@ -183,6 +189,7 @@ export const UserSchema = z.object({
   lti_user_id: z.string().nullable(),
   name: z.string().nullable(),
   stripe_customer_id: z.string().nullable(),
+  terms_accepted_at: DateFromISOString.nullable(),
   uid: z.string(),
   uin: z.string().nullable(),
   user_id: IdSchema,
@@ -240,6 +247,7 @@ export type Question = z.infer<typeof QuestionSchema>;
 
 export const WorkspaceSchema = z.object({
   created_at: DateFromISOString,
+  disk_usage_bytes: z.coerce.number().nullable(), // This is BIGINT, but always fits a number
   heartbeat_at: DateFromISOString.nullable(),
   hostname: z.string().nullable(),
   id: IdSchema,
@@ -359,9 +367,6 @@ export const EnrollmentSchema = z.object({
   course_instance_id: IdSchema,
   created_at: DateFromISOString,
   id: IdSchema,
-  // Currently unused.
-  // TODO: remove from schema entirely?
-  role: z.any(),
   user_id: IdSchema,
 });
 export type Enrollment = z.infer<typeof EnrollmentSchema>;
@@ -461,6 +466,7 @@ export const AssessmentQuestionSchema = z.object({
   max_submission_score_hist: z.array(z.number()).nullable(),
   max_submission_score_variance: z.number().nullable(),
   mean_question_score: z.number().nullable(),
+  median_question_score: z.number().nullable(),
   number: z.number().nullable(),
   number_in_alternative_group: z.number().nullable(),
   number_submissions_hist: z.array(z.number()).nullable(),
@@ -583,7 +589,7 @@ export const SubmissionSchema = z.object({
   grading_requested_at: DateFromISOString.nullable(),
   id: IdSchema,
   manual_rubric_grading_id: IdSchema.nullable(),
-  mode: z.enum(['Public', 'Exam', 'SEB']).nullable(),
+  mode: ModeSchema.nullable(),
   override_score: z.number().nullable(),
   params: z.record(z.string(), z.any()).nullable(),
   partial_scores: z.record(z.string(), z.any()).nullable(),
@@ -778,7 +784,7 @@ export const AssessmentInstanceSchema = z.object({
   last_client_fingerprint_id: IdSchema.nullable(),
   max_bonus_points: z.number().nullable(),
   max_points: z.number().nullable(),
-  mode: z.enum(['Public', 'Exam', 'SEB']).nullable(),
+  mode: ModeSchema.nullable(),
   modified_at: DateFromISOString,
   number: z.number().nullable(),
   open: z.boolean().nullable(),
@@ -881,6 +887,34 @@ export const CourseInstancePermissionSchema = z.object({
   id: IdSchema,
 });
 export type CourseInstancePermission = z.infer<typeof CourseInstancePermissionSchema>;
+
+export const AlternativeGroupSchema = z.object({
+  advance_score_perc: z.number().nullable(),
+  assessment_id: IdSchema,
+  id: IdSchema,
+  number: z.number().nullable(),
+  number_choose: z.number().nullable(),
+  zone_id: IdSchema,
+});
+export type AlternativeGroup = z.infer<typeof AlternativeGroupSchema>;
+
+export const ZoneSchema = z.object({
+  advance_score_perc: z.number().nullable(),
+  assessment_id: IdSchema,
+  best_questions: z.number().nullable(),
+  id: IdSchema,
+  max_points: z.number().nullable(),
+  number: z.number().nullable(),
+  number_choose: z.number().nullable(),
+  title: z.string().nullable(),
+});
+export type Zone = z.infer<typeof ZoneSchema>;
+
+export const AdministratorSchema = z.object({
+  id: IdSchema,
+  user_id: IdSchema,
+});
+export type Administrator = z.infer<typeof AdministratorSchema>;
 
 // Result of grading_job_status sproc
 export const GradingJobStatusSchema = z.enum([