+++ apps/prairielearn/src/middlewares/autoEnroll.ts
@@ -15,22 +15,16 @@ export default asyncHandler(async (req, res, next) => {
     res.locals.authz_data.authn_has_student_access &&
     !res.locals.authz_data.authn_has_student_access_with_enrollment
   ) {
-    const didEnroll = await ensureCheckedEnrollment({
+    await ensureCheckedEnrollment({
       institution: res.locals.institution,
       course: res.locals.course,
       course_instance: res.locals.course_instance,
       authz_data: res.locals.authz_data,
-      redirect: res.redirect.bind(res),
     });
 
-    if (didEnroll) {
-      // This is the only part of the `authz_data` that would change as a
-      // result of this enrollment, so we can just update it directly.
-      res.locals.authz_data.has_student_access_with_enrollment = true;
-    } else {
-      // The user has already been redirected to an appropriate page; do nothing.
-      return;
-    }
+    // This is the only part of the `authz_data` that would change as a
+    // result of this enrollment, so we can just update it directly.
+    res.locals.authz_data.has_student_access_with_enrollment = true;
   }
 
   next();