+++ apps/prairielearn/src/models/course.ts
@@ -17,6 +17,7 @@ const CourseWithPermissionsSchema = CourseSchema.extend({
     has_course_permission_preview: z.boolean(),
   }),
 });
+export type CourseWithPermissions = z.infer<typeof CourseWithPermissionsSchema>;
 
 export async function selectCourseById(course_id: string): Promise<Course> {
   return await queryRow(
@@ -40,9 +41,11 @@ export async function getCourseCommitHash(coursePath: string): Promise<string> {
     });
     return stdout.trim();
   } catch (err) {
-    throw error.makeWithData(`Could not get git status; exited with code ${err.code}`, {
-      stdout: err.stdout,
-      stderr: err.stderr,
+    throw new error.AugmentedError(`Could not get git status; exited with code ${err.code}`, {
+      data: {
+        stdout: err.stdout,
+        stderr: err.stderr,
+      },
     });
   }
 }