+++ apps/prairielearn/src/pages/instructorQuestions/instructorQuestions.ts
@@ -32,7 +32,7 @@ router.get(
     const courseDirExists = await fs.pathExists(res.locals.course.path);
     res.send(
       QuestionsPage({
-        questions: questions,
+        questions,
         course_instances: courseInstances,
         showAddQuestionButton:
           res.locals.authz_data.has_course_permission_edit &&
@@ -69,7 +69,7 @@ router.post('/', (req, res, next) => {
       });
     });
   } else {
-    next(error.make(400, `unknown __action: ${req.body.__action}`));
+    next(new error.HttpStatusError(400, `unknown __action: ${req.body.__action}`));
   }
 });