+++ packages/compiled-assets/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/compiled-assets
 
+## 2.0.17
+
+### Patch Changes
+
+- c7e6553: Upgrade all JavaScript dependencies
+- Updated dependencies [c7e6553]
+  - @prairielearn/html@3.1.7
+
+## 2.0.16
+
+### Patch Changes
+
+- 207602a: Upgrade all JavaScript dependencies
+- Updated dependencies [207602a]
+  - @prairielearn/html@3.1.6
+
 ## 2.0.15
 
 ### Patch Changes